fix: refactor the Login component into the write page.
