--- apps/minsu_dev/src/app/write/write.module.css
@@ -1,7 +1,0 @@
-.loginContainer {
-  display: flex;
-  flex-direction: column;
-  padding-top: 3rem;
-  padding-bottom: 5rem;
-  width: 100%;
-}
