+++ apps/minsu_dev/src/foo/login/login.module.css
@@ -17,3 +17,11 @@
 .errorMessage {
   color: #fff;
 }
+
+.loginContainer {
+  display: flex;
+  flex-direction: column;
+  padding-top: 3rem;
+  padding-bottom: 5rem;
+  width: 100%;
+}
